changed how final wpm/accuracy is calculated; scores are now verified by the server before saving

+++ init.sql
@@ -45,7 +45,7 @@ CREATE TABLE "results" (
     "user_id" INTEGER REFERENCES "users"("id"),
     "quote_id" INTEGER REFERENCES "quote"("id"),
     "wpm" INTEGER NOT NULL,
-    "accuracy" INTEGER NOT NULL,
+    "accuracy" DECIMAL NOT NULL,
     "played_at" TIMESTAMP DEFAULT NOW()
 );
